Finished cherry pick; cleaned up tests; merged upstream changes

+++ cmd/harbor/root/artifact/tags.go
@@ -50,7 +50,11 @@ func CreateTagsCmd() *cobra.Command {
 			var err error
 
 			if len(args) > 0 {
-				projectName, repoName, reference := utils.ParseProjectRepoReference(args[0])
+				projectName, repoName, reference, parseError := utils.ParseProjectRepoReference(args[0])
+				if parseError != nil {
+					log.Errorf("failed to parse project/repo/reference: %v", parseError)
+					return
+				}
 				tag := args[1]
 				err = api.CreateTag(projectName, repoName, reference, tag)
 			} else {
@@ -80,12 +84,16 @@ func ListTagsCmd() *cobra.Command {
 		Short:   "List tags of an artifact",
 		Example: `harbor artifact tags list <project>/<repository>/<reference>`,
 		Run: func(cmd *cobra.Command, args []string) {
-			var err error
+			var err, parseError error
 			var tags *artifact.ListTagsOK
 			var projectName, repoName, reference string
 
 			if len(args) > 0 {
-				projectName, repoName, reference = utils.ParseProjectRepoReference(args[0])
+				projectName, repoName, reference, parseError = utils.ParseProjectRepoReference(args[0])
+				if parseError != nil {
+					log.Errorf("failed to parse project/repo/reference: %v", parseError)
+					return
+				}
 			} else {
 				projectName, err = prompt.GetProjectNameFromUser()
 				if err != nil {
@@ -127,7 +135,11 @@ func DeleteTagsCmd() *cobra.Command {
 			var err error
 
 			if len(args) > 0 {
-				projectName, repoName, reference := utils.ParseProjectRepoReference(args[0])
+				projectName, repoName, reference, parseError := utils.ParseProjectRepoReference(args[0])
+				if parseError != nil {
+					log.Errorf("failed to parse project/repo/reference: %v", parseError)
+					return
+				}
 				tag := args[1]
 				err = api.DeleteTag(projectName, repoName, reference, tag)
 			} else {
